# docs/swagger.yaml
...
info:
  contact: {}
paths:
  /api//sexnovel/chapte/list:
    get:
      parameters:
      - in: query
        name: id
        type: integer
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: pid
        type: integer
      - in: query
        name: title
        type: string
      responses: {}
      summary: 小说章节列表
      tags:
      - 小说
...
  /api/blog/comment/list:
    get:
      parameters:
      - in: query
        name: appid
        type: integer
      - in: query
        name: blog_id
        type: integer
      - in: query
        name: id
        type: integer
      - in: query
        name: ip
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: order
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: parent_id
        type: integer
      - in: query
        name: seximg_id
        type: integer
      - in: query
        name: type
        type: integer
      - in: query
        name: userid
        type: string
      - in: query
        name: vlog_id
        type: integer
      - in: query
        name: vod_id
        type: integer
      responses: {}
      summary: 获取视频评论
...
  /api/seximg/list:
    get:
      parameters:
      - in: query
        name: after_create_time
        type: string
      - in: query
        name: appId
        type: integer
      - in: query
        name: id
        type: integer
      - in: query
        name: limit
        type: integer
      - in: query
        name: order
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: status
        type: integer
      - in: query
        name: time
        type: string
      - in: query
        name: title
        type: string
      - in: query
        name: top
        type: integer
      - in: query
        name: typeid
        type: integer
      - in: query
        name: wd
        type: string
      responses: {}
      summary: 色图列表
      tags:
      - 色图
  /api/seximgtype/list:
    get:
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      responses: {}
      summary: 色图分类列表
      tags:
      - 色图
...
  /api/sexnovel/list:
    get:
      parameters:
      - in: query
        name: appId
        type: integer
      - in: query
        name: id
        type: integer
      - in: query
        name: labelsId
        type: integer
      - in: query
        name: limit
        type: integer
      - in: query
        name: order
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: status
        type: integer
      - in: query
        name: title
        type: string
      - in: query
        name: top
        type: integer
      - in: query
        name: typeid
        type: integer
      responses: {}
      summary: 小说列表
      tags:
      - 小说
...
  /api/user/blog/star/list:
    get:
      consumes:
      - application/json
      description: 获取某个用户的点赞记录
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: type
        type: integer
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 获取点赞记录
...
  /api/user/list:
    get:
      consumes:
      - application/json
      description: 返回用户信息
      parameters:
      - in: query
        name: denycomment
        type: integer
      - in: query
        name: id
        type: integer
      - in: query
        name: ip
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: name
        type: string
      - in: query
        name: order
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: phone
        type: string
      - in: query
        name: type
        type: string
      responses: {}
      summary: 搜索用户列表
      tags:
      - 用户
...
  /api/user/seximg/star/list:
    get:
      consumes:
      - application/json
      description: 获取某个用户的点赞收藏记录
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: type
        type: integer
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 获取点赞收藏记录
      tags:
      - 色图
...
  /api/user/sexnovel/history/list:
    get:
      consumes:
      - application/json
      description: 创建小说观看记录，单用户与单小说唯一。
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 获取小说观看记录
      tags:
      - 小说
...
  /api/user/sexnovel/star/list:
    get:
      consumes:
      - application/json
      description: 获取某个用户的点赞收藏记录
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: type
        type: integer
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 获取点赞收藏记录
      tags:
      - 小说
...
  /api/user/vod/history/list:
    get:
      consumes:
      - application/json
      description: 创建视频播放记录，单用户与单视频唯一。
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 获取播放记录
      tags:
      - 长视频
...
  /api/user/vod/star/list:
    get:
      consumes:
      - application/json
      description: 获取某个用户的收藏列表
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: 获取收藏记录
      tags:
      - 长视频
  /api/vip/list:
    get:
      consumes:
      - application/json
      description: 返回VIP列表数据
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: onStatus
        type: boolean
      - in: query
        name: order
        type: string
      - in: query
        name: page
        type: integer
      responses: {}
      summary: VIP列表数据
      tags:
      - VIP
...
  /api/vlog/list:
    get:
      consumes:
      - application/json
      description: 返回短视频列表
      parameters:
      - in: query
        name: after_create_time
        type: string
      - in: query
        name: appId
        type: integer
      - in: query
        name: detail
        type: string
      - in: query
        name: id
        type: integer
      - in: query
        name: limit
        type: integer
      - in: query
        name: matchid
        type: integer
      - in: query
        name: order
        type: string
      - in: query
        name: page
        type: integer
      - description: 是否自己发布的接口区分参数 多查几个状态 0 1
        in: query
        name: self
        type: integer
      - in: query
        name: star
        type: integer
      - in: query
        name: status
        type: integer
      - in: query
        name: time
        type: string
      - in: query
        name: top
        type: integer
      - in: query
        name: topicid
        type: integer
      - in: query
        name: type
        type: string
      - in: query
        name: user_id
        type: string
      - in: query
        name: userput
        type: integer
      - in: query
        name: wd
        type: string
      responses: {}
      summary: 短视频列表
      tags:
      - 短视频
...
  /api/vod/list:
    get:
      consumes:
      - application/json
      description: 返回视频列表数据
      parameters:
      - in: query
        name: after_create_time
        type: string
      - in: query
        name: appId
        type: integer
      - in: query
        name: id
        type: integer
      - in: query
        name: job_status
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: labels
        type: array
      - in: query
        name: limit
        type: integer
      - description: 番号
        in: query
        name: number
        type: string
      - in: query
        name: order
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: status
        type: integer
      - in: query
        name: title
        type: string
      - in: query
        name: topicid
        type: integer
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: types
        type: array
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: users
        type: array
      responses: {}
      summary: 视频列表
      tags:
      - 长视频
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"